refactor random augmentation methods to a better place
It is more natural to have the extractor know and handel augmentaion.
